LayoutTests/imported/w3c/web-platform-tests/css/css-align/content-distribution/parse-align-content-001-expected.txt


PASS Checking align-content: normal 
FAIL Checking align-content: start assert_equals: align-content computed style is not what is should. expected "start" but got "normal"
FAIL Checking align-content: end assert_equals: align-content computed style is not what is should. expected "end" but got "normal"
FAIL Checking align-content: center assert_equals: align-content computed style is not what is should. expected "center" but got "normal"
FAIL Checking align-content: flex-start assert_equals: align-content computed style is not what is should. expected "flex-start" but got "normal"
FAIL Checking align-content: flex-end assert_equals: align-content computed style is not what is should. expected "flex-end" but got "normal"
FAIL Checking align-content: stretch assert_equals: align-content computed style is not what is should. expected "stretch" but got "normal"
FAIL Checking align-content: space-around assert_equals: align-content computed style is not what is should. expected "space-around" but got "normal"
FAIL Checking align-content: space-between assert_equals: align-content computed style is not what is should. expected "space-between" but got "normal"
FAIL Checking align-content: space-evenly assert_equals: align-content computed style is not what is should. expected "space-evenly" but got "normal"
FAIL Checking align-content: baseline assert_equals: align-content computed style is not what is should. expected "first baseline" but got "normal"
FAIL Checking align-content: first baseline assert_equals: align-content computed style is not what is should. expected "first baseline" but got "normal"
FAIL Checking align-content: last baseline assert_equals: align-content computed style is not what is should. expected "last baseline" but got "normal"
FAIL Checking align-content: safe flex-end assert_equals: align-content computed style is not what is should. expected "safe flex-end" but got "normal"
FAIL Checking align-content: unsafe end assert_equals: align-content computed style is not what is should. expected "unsafe end" but got "normal"
FAIL Checking align-content: safe end assert_equals: align-content computed style is not what is should. expected "safe end" but got "normal"
FAIL Checking align-content: unsafe flex-start assert_equals: align-content computed style is not what is should. expected "unsafe flex-start" but got "normal"
FAIL Checking align-content: safe center assert_equals: align-content computed style is not what is should. expected "safe center" but got "normal"
